Register pusher, implement `SocketID`
