+++ Equality.Core/Http/Interfaces/IWebsocketClient.cs
@@ -5,6 +5,8 @@ namespace Equality.Http
 {
     public interface IWebsocketClient
     {
+        public string SocketID { get; }
+
         public Task BindEventAsync(string channelName, string eventName, Action<string> listener);
 
         public void UnbindEvent(string channelName, string eventName);
