+++ Equality/App.xaml.cs
@@ -59,8 +59,7 @@ namespace Equality
 
             var serviceLocator = ServiceLocator.Default;
 
-            RegisterPusher(serviceLocator);
-            serviceLocator.RegisterType<ITokenResolverService, TokenResolver>();
+            serviceLocator.RegisterType<ITokenResolver, TokenResolver>();
             serviceLocator.RegisterType<INotificationService, NotificationService>();
             serviceLocator.RegisterTypeAndInstantiate<ExceptionHandler>();
 
@@ -116,39 +115,6 @@ namespace Equality
             Log.Info("Calling base.OnStartup");
 
             base.OnStartup(e);
-        }
-
-        private void RegisterPusher(IServiceLocator serviceLocator)
-        {
-            var client = new Http.PusherClient("7c6a91460be1e040ce8c", new PusherOptions
-            {
-                Cluster = "eu",
-                Encrypted = true,
-            });
-
-            void HandleError(object sender, PusherException error)
-            {
-                if ((int)error.PusherCode < 5000) {
-                    // Error recevied from Pusher cluster, use PusherCode to filter.
-                } else {
-                    if (error is ChannelUnauthorizedException unauthorizedAccess) {
-                        // Private and Presence channel failed authorization with Forbidden (403)
-                    } else if (error is ChannelAuthorizationFailureException httpError) {
-                        // Authorization endpoint returned an HTTP error other than Forbidden (403)
-                    } else if (error is OperationTimeoutException timeoutError) {
-                        // A client operation has timed-out. Governed by PusherOptions.ClientTimeout
-                    } else if (error is ChannelDecryptionException decryptionError) {
-                        // Failed to decrypt the data for a private encrypted channel
-                    } else {
-                        // Handle other errors
-                    }
-                }
-
-                Trace.TraceError($"{error}");
-            }
-            client.Error += HandleError;
-
-            serviceLocator.RegisterInstance<IWebsocketClient>(client);
         }
     }
 }
