+++ Equality/ViewModels/LoadingWindowViewModel.cs
@@ -1,12 +1,16 @@
 ﻿using System;
+using System.Diagnostics;
 using System.Threading.Tasks;
 
+using Catel.IoC;
 using Catel.Services;
 
 using Equality.Data;
 using Equality.Http;
 using Equality.MVVM;
 using Equality.Services;
+
+using PusherClient;
 
 namespace Equality.ViewModels
 {
@@ -32,6 +36,7 @@ namespace Equality.ViewModels
                 bool result = await IsValidToken(apiToken);
 
                 if (result) {
+                    RegisterPusher();
                     OpenMainPage();
                 } else {
                     OpenAuthorizationPage();
@@ -68,6 +73,43 @@ namespace Equality.ViewModels
 
                 throw;
             }
+        }
+
+        protected void RegisterPusher()
+        {
+            var client = new Http.PusherClient("7c6a91460be1e040ce8c", new PusherOptions
+            {
+                Authorizer = new HttpAuthorizer("http://equality/broadcasting/auth")
+                {
+                    AuthenticationHeader = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", StateManager.ApiToken)
+                },
+                Cluster = "eu",
+                Encrypted = true,
+            });
+
+            void HandleError(object sender, PusherException error)
+            {
+                if ((int)error.PusherCode < 5000) {
+                    // Error recevied from Pusher cluster, use PusherCode to filter.
+                } else {
+                    if (error is ChannelUnauthorizedException unauthorizedAccess) {
+                        // Private and Presence channel failed authorization with Forbidden (403)
+                    } else if (error is ChannelAuthorizationFailureException httpError) {
+                        // Authorization endpoint returned an HTTP error other than Forbidden (403)
+                    } else if (error is OperationTimeoutException timeoutError) {
+                        // A client operation has timed-out. Governed by PusherOptions.ClientTimeout
+                    } else if (error is ChannelDecryptionException decryptionError) {
+                        // Failed to decrypt the data for a private encrypted channel
+                    } else {
+                        // Handle other errors
+                    }
+                }
+
+                Trace.TraceError($"{error}");
+            }
+            client.Error += HandleError;
+
+            ServiceLocator.Default.RegisterInstance<IWebsocketClient>(client);
         }
 
         protected void OpenMainPage()
